+++ indicstemmer/templates/indicstemmer.html
@@ -65,6 +65,7 @@
             <select id="lang" name="lang" class="form-control">
                 <option value="ml_IN">Malayalam</option>
                 <option value="hi_IN">Hindi</option>                
+                <option value="pu_IN">Punjabi</option>                
             </select>
 	</div>
            <input type="submit" id="Stem_button" value="Stem"  name="action" class="btn btn-default pull-right"/>
